moves audio-wave graphing to own package to enable quick excluding from build
